Crank up the warnings level still further.
We are now requesting essentially all of the warnings that GCC can
produce.  Many of the code changes required are confirming that C's
somewhat confusing type conversion rules do in fact do the Right Thing
in each case.

+++ alg-md5.c
@@ -95,19 +95,19 @@ md5_process_bytes (const void *buffer, size_t len, struct md5_ctx *ctx)
      both inputs first.  */
   if (ctx->buflen != 0)
     {
-      size_t left_over = ctx->buflen;
-      size_t add = 128 - left_over > len ? len : 128 - left_over;
+      uint32_t left_over = ctx->buflen;
+      uint32_t add = 128 - left_over > len ? (uint32_t)len : 128 - left_over;
 
       memcpy (&ctx->buffer[left_over], buffer, add);
       ctx->buflen += add;
 
       if (ctx->buflen > 64)
         {
-          md5_process_block (ctx->buffer, ctx->buflen & ~63, ctx);
+          md5_process_block (ctx->buffer, ctx->buflen & ~63u, ctx);
 
           ctx->buflen &= 63;
           /* The regions in the following copy operation cannot overlap.  */
-          memcpy (ctx->buffer, &ctx->buffer[(left_over + add) & ~63],
+          memcpy (ctx->buffer, &ctx->buffer[(left_over + add) & ~63u],
                   ctx->buflen);
         }
 
@@ -118,18 +118,18 @@ md5_process_bytes (const void *buffer, size_t len, struct md5_ctx *ctx)
   /* Process available complete blocks.  */
   if (len > 64)
     {
-      md5_process_block (buffer, len & ~63, ctx);
-      buffer = (const char *) buffer + (len & ~63);
+      md5_process_block (buffer, len & ~63u, ctx);
+      buffer = (const char *) buffer + (len & ~63u);
       len &= 63;
     }
 
   /* Move remaining bytes in internal buffer.  */
   if (len > 0)
     {
-      size_t left_over = ctx->buflen;
+      uint32_t left_over = ctx->buflen;
 
       memcpy (&ctx->buffer[left_over], buffer, len);
-      left_over += len;
+      left_over += (uint32_t)len;
       if (left_over >= 64)
         {
           md5_process_block (ctx->buffer, 64, ctx);
